Tatsuhiro Nishioka:
I changed the <stall> tag for sharper stall.
That may not be the correct one but can be closer.

Taxing with this aircraft is a bit hard because its tail wheel
doesn't rotate as rudder moves.
dfric and sfric of the main gears are set to 0.6 and 0.7 respectively
but these don't effect so much in taxing.
So I added the steer control for the tail wheel, I don't know if
Hayate has this type of tail wheel though.